pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "egen-iam-local-backend"
version = "1.0.0"
description = "EGEN National IAM Local (Module 02) - Gestion des identités et accès locaux par établissement, rôles, permissions et habilitations"
requires-python = ">=3.11"
authors = [
    {name = "EGEN Team", email = "team@egen.gouv.ga"}
]
readme = "README.md"
license = {file = "LICENSE"}
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Education",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Framework :: FastAPI",
    "Topic :: Internet :: WWW/HTTP :: HTTP Servers",
    "Topic :: Security",
]
keywords = ["egen", "iam", "identity", "access", "management", "gabon", "backend", "fastapi", "permissions"]

dependencies = [
    # Web framework
    "fastapi==0.111.0",
    "uvicorn[standard]==0.30.1",

    # Validation & settings
    "pydantic==2.7.1",
    "pydantic-settings==2.3.0",
    "python-dotenv==1.0.1",

    # Base de données
    "sqlalchemy[asyncio]==2.0.30",
    "asyncpg==0.29.0",
    "alembic==1.13.1",

    # Cache
    "redis==5.0.4",

    # Kafka
    "confluent-kafka==2.4.0",

    # JWT
    "PyJWT==2.8.0",

    # HTTP client
    "httpx==0.27.0",

    # File uploads
    "python-multipart==0.0.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "httpx>=0.27.0",
    "faker>=24.0.0",
    "ruff>=0.4.0",
]

test = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "httpx>=0.27.0",
    "faker>=24.0.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["app/tests"]

[project.urls]
Homepage = "https://github.com/egen-gabon/egen-national-backend-iam-local"
Repository = "https://github.com/egen-gabon/egen-national-backend-iam-local"
Documentation = "https://github.com/egen-gabon/egen-national-backend-iam-local/docs"

[project.scripts]
egen-iam-local = "app.main:main"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
target-version = "py311"
line-length = 88
select = ["E", "F", "W", "I", "N", "B", "C90", "UP"]
ignore = ["E501", "B008"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.isort]
profile = "black"
line_length = 88
